pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "e-commerce-test-suite"
version = "1.0.0"
description = "Enterprise-grade full-stack test automation framework with 300+ tests"
authors = [
    {name = "Jason Teixeira", email = "jason@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = [
    "testing",
    "automation",
    "qa",
    "playwright",
    "selenium",
    "api-testing",
    "performance-testing",
    "security-testing",
    "accessibility-testing",
    "e2e-testing"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "pytest>=7.4.3",
    "playwright>=1.41.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "isort>=5.13.0",
    "pylint>=3.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]
test = [
    "pytest-xdist>=3.5.0",
    "pytest-cov>=4.1.0",
    "pytest-html>=4.1.0",
    "pytest-timeout>=2.2.0",
    "pytest-rerunfailures>=13.0",
    "allure-pytest>=2.13.2",
]

[project.urls]
Homepage = "https://github.com/JasonTeixeira/E-Commerce-Test-Suite"
Repository = "https://github.com/JasonTeixeira/E-Commerce-Test-Suite"
Issues = "https://github.com/JasonTeixeira/E-Commerce-Test-Suite/issues"

# ========================================
# Black Configuration (Code Formatting)
# ========================================
[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | reports
)/
'''

# ========================================
# isort Configuration (Import Sorting)
# ========================================
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = [
    "venv/*",
    ".venv/*",
    "reports/*",
    "logs/*"
]

# ========================================
# Pytest Configuration
# ========================================
[tool.pytest.ini_options]
minversion = "7.4"
testpaths = [
    "ui_tests",
    "api_tests",
    "integration",
    "accessibility",
    "security"
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "--tb=short",
    "--disable-warnings",
    "-v",
    "--color=yes",
    "--code-highlight=yes",
]

# Test markers
markers = [
    "smoke: Quick smoke tests for critical functionality",
    "regression: Full regression test suite",
    "critical: Critical path tests",
    "ui: User interface tests",
    "api: API tests",
    "performance: Performance and load tests",
    "security: Security vulnerability tests",
    "accessibility: Accessibility compliance tests",
    "visual: Visual regression tests",
    "integration: Integration and E2E tests",
    "slow: Tests that take longer to run",
    "flaky: Tests that may occasionally fail",
    "skip_ci: Tests to skip in CI environment",
    "browser_chromium: Tests for Chromium browser",
    "browser_firefox: Tests for Firefox browser",
    "browser_webkit: Tests for WebKit browser",
    "mobile: Mobile responsive tests",
    "desktop: Desktop-only tests",
]

# Timeout configuration
timeout = 300
timeout_method = "thread"

# Logging
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_file = "logs/pytest.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s [%(levelname)8s] [%(filename)s:%(lineno)d] %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

# Coverage
[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/venv/*",
    "*/.venv/*",
    "*/conftest.py",
    "*/setup.py",
]
branch = true
parallel = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.coverage.html]
directory = "reports/coverage"

# ========================================
# MyPy Configuration (Type Checking)
# ========================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
pretty = true
show_error_codes = true
show_error_context = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "playwright.*",
    "allure.*",
    "pytest_html.*",
    "locust.*",
]
ignore_missing_imports = true

# ========================================
# Pylint Configuration
# ========================================
[tool.pylint.main]
py-version = "3.11"
jobs = 0
limit-inference-results = 100
suggestion-mode = true
fail-under = 8.0

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
    "W0212",  # protected-access (needed for testing)
    "C0415",  # import-outside-toplevel
]

[tool.pylint.format]
max-line-length = 100
max-module-lines = 1000

[tool.pylint.basic]
good-names = ["i", "j", "k", "ex", "Run", "_", "id", "db"]

[tool.pylint.design]
max-args = 10
max-attributes = 15
max-bool-expr = 5
max-branches = 15
max-locals = 20
max-parents = 7
max-public-methods = 25
max-returns = 6
max-statements = 50
min-public-methods = 1

# ========================================
# Bandit Configuration (Security Linting)
# ========================================
[tool.bandit]
exclude_dirs = [
    "/tests/",
    "/venv/",
    "/.venv/",
]
skips = ["B101", "B601"]
